move py script
